Support added for Date, Time and DateTime class objects.
Changes:

- Added `formatted_time` formating support for Date class objects, e.g. Date.today.frmt_dd_mm_yy.
- Added `formatted_time` formating support for DateTime class objects, e.g. DateTime.now.frmt_dd_mm_yy.
- Added `formatted_time` formating support for Time class objects, e.g. Time.now.frmt_dd_mm_yy.
- Default seperator changed from `:` to `/`
